extract zip code to ziputils
